requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pycparser==2.20
pythonnet==2.5.2
pywebview==3.4
Werkzeug==2.0.0
